# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Case Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegistrationBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Duplicate username

  # ── Users & Assignments ───────────────────────────────────────────────────
...
  /projects/{projectId}/bugs:
    get:
      operationId: listBugs
      tags: [bugs]
      summary: List all bugs for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: developerId
          in: query
          required: false
          schema:
            type: integer
        - name: ticketNumber
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of bugs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bug"
  /bugs/{bugId}:
    get:
      operationId: getBug
      tags: [bugs]
      summary: Get a single bug with full detail
      parameters:
        - name: bugId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bug detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{bugId}/assign:
    patch:
      operationId: assignBug
      tags: [bugs]
      summary: Assign a bug to a developer
      parameters:
        - name: bugId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignBugBody"
      responses:
        "200":
          description: Bug assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{bugId}/status:
    patch:
      operationId: updateBugStatus
      tags: [bugs]
      summary: Update bug status
      parameters:
        - name: bugId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBugStatusBody"
      responses:
        "200":
          description: Bug status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{bugId}/notes:
    patch:
      operationId: updateBugNotes
      tags: [bugs]
      summary: Update developer notes on a bug
      parameters:
        - name: bugId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBugNotesBody"
      responses:
        "200":
          description: Bug notes updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{bugId}/reassign:
    patch:
      operationId: reassignBug
      tags: [bugs]
      summary: Reassign a bug to another developer
      parameters:
        - name: bugId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignBugBody"
      responses:
        "200":
          description: Bug reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"

  # ── Team Discussions ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    TestCase:
      type: object
      properties:
        id:
          type: integer
        useCaseId:
          type: integer
        caseNumber:
          type: integer
        title:
          type: string
        createdAt:
          type: string
      required: [id, useCaseId, caseNumber, title, createdAt]
...
    TestExecution:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testRunId:
          type: integer
          nullable: true
        iterationNumber:
          type: integer
        testerName:
          type: string
        status:
          type: string
        executedAt:
          type: string
        stepResults:
          type: array
          items:
            $ref: "#/components/schemas/StepResult"
      required: [id, testCaseId, iterationNumber, testerName, status, executedAt]
...
    StepResult:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        stepId:
          type: integer
        actualResult:
          type: string
          nullable: true
        comments:
          type: string
          nullable: true
        passed:
          type: boolean
          nullable: true
        recordedAt:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required: [id, executionId, stepId, recordedAt, attachments]
...
    Attachment:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        field:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
      required: [id, entityType, entityId, field, fileName, fileUrl, fileType, createdAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, username, name, role, isActive, createdAt]
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    CreateRegistrationBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [ADMIN, AUTHOR, USER]
      required: [username, password, name, role]
...
    Defect:
      type: object
      properties:
        id:
          type: integer
        testRunId:
          type: integer
        testCaseId:
          type: integer
        executionId:
          type: integer
        testerNotes:
          type: string
          nullable: true
        status:
          type: string
        retestReason:
          type: string
          nullable: true
        acceptedByBusinessNote:
          type: string
          nullable: true
        rejectionLog:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        testCase:
          $ref: "#/components/schemas/TestCase"
          nullable: true
        execution:
          $ref: "#/components/schemas/TestExecution"
          nullable: true
        notes:
          type: array
          items:
            $ref: "#/components/schemas/DefectNote"
      required: [id, testRunId, testCaseId, executionId, status, createdAt, updatedAt]
...
    DefectNote:
      type: object
      properties:
        id:
          type: integer
        defectId:
          type: integer
        discussionId:
          type: integer
          nullable: true
        addedByUserId:
          type: integer
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        addedByUser:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, defectId, addedByUserId, note, createdAt]
    StatusAuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        changedByUserId:
          type: integer
        fromStatus:
          type: string
        toStatus:
          type: string
        reason:
          type: string
          nullable: true
        changedAt:
          type: string
          format: date-time
        changedByUser:
          $ref: "#/components/schemas/User"
          nullable: true
      required: [id, entityType, entityId, changedByUserId, fromStatus, toStatus, changedAt]

    # ── Bug Schemas ──────────────────────────────────────────────────────────
    Bug:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        defectId:
          type: integer
        bugNumber:
          type: integer
        supportTicketNumber:
          type: string
          nullable: true
        assignedDeveloperId:
          type: integer
          nullable: true
        status:
          type: string
        developerNotes:
          type: string
          nullable: true
        failedToResolveReason:
          type: string
          nullable: true
        openedAt:
          type: string
          format: date-time
        assignedAt:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        testAt:
          type: string
          format: date-time
          nullable: true
        failedToResolveAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        defect:
          $ref: "#/components/schemas/Defect"
          nullable: true
        assignedDeveloper:
          $ref: "#/components/schemas/User"
          nullable: true
        auditLog:
          type: array
          items:
            $ref: "#/components/schemas/StatusAuditLogEntry"
      required: [id, projectId, defectId, bugNumber, status, openedAt, createdAt, updatedAt]
    AssignBugBody:
      type: object
      properties:
        developerId:
          type: integer
        supportTicketNumber:
          type: string
      required: [developerId]
    UpdateBugStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [OPEN, ASSIGNED, RESOLVED, TEST, FAILED_TO_RESOLVE, CLOSED]
        reason:
          type: string
      required: [status]
    UpdateBugNotesBody:
      type: object
      properties:
        notes:
          type: string
      required: [notes]
    ReassignBugBody:
      type: object
      properties:
        developerId:
          type: integer
      required: [developerId]

    # ── Team Discussion Schemas ──────────────────────────────────────────────
...