# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UAT Test Case Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  name: string;
  email?: string | null;
  role: string;
  isActive: boolean;
  createdAt: string;
}